+++ src/actions/discussion.js
@@ -1,7 +1,7 @@
+import { dispatchEntities } from '../utils/helpers';
 import { discussionSchema } from '../models/schema';
 import { ADD_DISCUSSION_TO_SUBJECT } from './actionTypes';
-import { dispatchEntities } from '../components/utils/helpers';
-import { requestGraphql } from '../components/utils/HTTPClient';
+import { requestGraphql } from '../utils/HTTPClient';
 import { mutationCreateDiscussion } from '../queryGenerators/discussionMutations';
 
 export const createDiscussion = input => dispatch => {
