+++ src/components/Grade/GradeDetails/SubjectsGrid.js
@@ -1,12 +1,12 @@
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import React, { Component } from 'react';
-import GridCard from '../../utils/GridCard';
+import GridCard from '../../shared/GridCard';
 import { withRouter } from 'react-router-dom';
-import GridButton from '../../utils/GridButton';
+import GridButton from '../../shared/GridButton';
 import AddCircleIcon from '@material-ui/icons/AddCircle';
-import SubjectCreateDialog from '../../Institution/SubjectCreateDialog';
-import { GridContainer, ResponsiveSubTitle } from '../../utils/SharedComponents';
+import CreateSubjectDialog from '../../Institution/CreateSubjectDialog';
+import { GridContainer, ResponsiveSubTitle } from '../../shared/SharedComponents';
 
 const Container = styled.div`
   display: flex;
@@ -45,7 +45,7 @@ class SubjectsGrid extends Component {
     return (
       <Container>
         <ResponsiveSubTitle>Disciplinas</ResponsiveSubTitle>
-        <SubjectCreateDialog
+        <CreateSubjectDialog
           gradeId={gradeId}
           open={subjectDialogOpen}
           onClose={this.handleSubjectDialogToggle}
