+++ src/components/Institution/EntryCode/AssociateDialog.js
@@ -5,8 +5,8 @@ import { connect } from 'react-redux';
 import { withFormik, Form } from 'formik';
 import { capitalize } from '@material-ui/core/utils/helpers';
 import { joinInstitution } from '../../../actions/institution';
-import { defaultDialogBreakpoints } from '../../utils/helpers';
-import { DefaultDialogTransition } from '../../utils/SharedComponents';
+import { defaultDialogBreakpoints } from '../../../utils/helpers';
+import { DefaultDialogTransition } from '../../shared/SharedComponents';
 import {
   Zoom,
   Input,
