+++ src/components/Institution/InstitutionDetails/EditInstitutionDialog.js
@@ -3,12 +3,12 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import React, { Component } from 'react';
 import { Form, withFormik } from 'formik';
-import ProgressButton from '../../utils/ProgressButton';
+import ProgressButton from '../../shared/ProgressButton';
 import { capitalize } from '@material-ui/core/utils/helpers';
-import { defaultDialogBreakpoints } from '../../utils/helpers';
 import { updateInstitution } from '../../../actions/institution';
+import { defaultDialogBreakpoints } from '../../../utils/helpers';
 import { NO_INSTUTION_SELECTED } from '../../../reducers/institution';
-import { DefaultDialogTransition } from '../../utils/SharedComponents';
+import { DefaultDialogTransition } from '../../shared/SharedComponents';
 import {
   Button,
   Dialog,
