+++ src/components/Subject/SubjectDetails/AddStudentDialog.js
@@ -4,11 +4,11 @@ import { connect } from 'react-redux';
 import React, { Fragment } from 'react';
 import { Form, withFormik } from 'formik';
 import AddTeacherEmptyView from './AddTeacherEmptyView';
-import ProgressButton from '../../utils/ProgressButton';
+import ProgressButton from '../../shared/ProgressButton';
 import { capitalize } from '@material-ui/core/utils/helpers';
-import { defaultDialogBreakpoints } from '../../utils/helpers';
 import { addStudentToSubject } from '../../../actions/subject';
-import { DefaultDialogTransition } from '../../utils/SharedComponents';
+import { defaultDialogBreakpoints } from '../../../utils/helpers';
+import { DefaultDialogTransition } from '../../shared/SharedComponents';
 import {
   Zoom,
   Input,
