+++ src/reducers/discussion.js
@@ -1,4 +1,4 @@
-import { concatIdIfNotContain } from '../components/utils/helpers';
+import { concatIdIfNotContain } from '../utils/helpers';
 import { SAVE_DISCUSSION, ADD_REPLY_TO_DISCUSSION } from '../actions/actionTypes';
 
 const DEFAULT_STATE = {
