+++ src/reducers/user.js
@@ -1,4 +1,4 @@
-import { concatIdIfNotContain } from '../components/utils/helpers';
+import { concatIdIfNotContain } from '../utils/helpers';
 import { SAVE_USER, SET_LOGGED_USER, REMOVE_USER, REMOVE_ALL_USERS } from '../actions/actionTypes';
 
 export const NO_USER_LOGGED = '-333';
